Added rudimentary block selector

+++ Assets/Backend/World/Block.cs
@@ -7,14 +7,14 @@ namespace BlockGame.Backend
     public struct Block
     {
         public string Name { get; set; }
-        public int BlockId { get; private set; }
+        public short BlockId { get; private set; }
         public bool IsOpaque { get; set; }
         public bool IsSolid { get; set; }
         public Color ReprColor { get; set; }
         public TextureCoords TexCoords { get; set; }
         
 
-        public Block (int blockId, string name = "NewBlock")
+        public Block (short blockId, string name = "NewBlock")
         {
             Name = name;
             BlockId = blockId;
